some test were added

+++ src/main/java/com/kraievskyi/lection1/model/Ball.java
@@ -8,11 +8,12 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @EqualsAndHashCode
 public class Ball extends Figure {
+
     private int radius;
 
     @Override
-    public double calculateVolume() {
-        super.setVolume((4.0 / 3.0) * Math.PI * (Math.pow(radius, 3)));
+    public int calculateVolume() {
+        super.setVolume((int) ((4.0 / 3.0) * Math.PI * (Math.pow(radius, 3))));
         return super.getVolume();
     }
